[user-sessions][l] Create user model
