--- views/user/login-failure.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="container">
-  <p>Login Failed</p>
-</div>
-{% endblock %}
