--- views/user/login-success.html
@@ -1,4 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<p>Welcome, {{ user.name }}!</p>
-{% endblock %}
